deeptier/requirement.txt
amqp==5.2.0
ansicon==1.89.0
arrow==1.3.0
asgiref==3.7.2
billiard==4.2.0
blessed==1.20.0
certifi==2023.11.17
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Django==4.2.7
django-appconf==1.0.6
django-cors-headers==4.3.1
django-phonenumber-field==7.2.0
django-picklefield==3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
et-xmlfile==1.1.0
idna==3.6
jinxed==1.2.1
kombu==5.3.5
Markdown==3.5.1
openpyxl==3.1.2
pillow==10.2.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2023.3.post1
rcssmin==1.1.1
redis==3.5.3
reportlab==4.0.8
requests==2.31.0
rjsmin==1.2.1
six==1.16.0
sqlparse==0.4.4
types-python-dateutil==2.8.19.20240106
typing_extensions==4.10.0
tzdata==2023.3
urllib3==2.1.0
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.6.0
gunicorn==21.2.0
setuptools==69.5.1